Defined THPS1/2 flag constants and added debug names loading.
The flag variable names were found in the THPS2 beta source code snippets.
NOTE: The list is still incomplete.

When loading a psx file it will look for a "DebugNames.dbg" file in the current directory. Normally only the THPS2x files include object names. But a lookup table can be generated so that we can still get the object names when loading the pc/dc/ps1 files.

I also included the crc32b calculation function. So the dbg file can contain checksum+name pairs, or just the name and the checksum will be generated for you.